hw1-data-collection/requirements.txt
datasets>=2.14.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
nltk>=3.8.0
nbformat>=5.9.0
nbconvert>=7.0.0
